+++ Income.h
@@ -10,7 +10,7 @@ class Income : public DataOfBalance {
     int incomeId;
 
 public:
-    Income(int incomeId = 0, int userId = 0, string date = "", string item = "", int amount = 0) {
+    Income(int incomeId = 0, int userId = 0, string date = "", string item = "", double amount = 0) {
         this->incomeId = incomeId;
         this->userId = userId;
         this->date = date;
